# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend server for VietCard application",
        "title": "VietCard Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "hynduf@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/card/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create New Card",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "card"
                ],
                "summary": "Create New Card",
                "parameters": [
                    {
                        "description": "Create Card Request",
                        "name": "create_card_request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateCardRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateCardResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/deck/copy": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Copy Deck",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "deck"
                ],
                "summary": "Copy Deck",
                "parameters": [
                    {
                        "description": "Copy Deck Request",
                        "name": "copy_deck_request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CopyDeckRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CopyDeckResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.CopyDeckRequest": {
            "type": "object",
            "required": [
                "deck_id"
            ],
            "properties": {
                "deck_id": {
                    "type": "string"
                }
            }
        },
        "dto.CopyDeckResponse": {
            "type": "object",
            "properties": {
                "deck": {
                    "$ref": "#/definitions/entity.DeckWithCards"
                },
                "deck_review": {
                    "$ref": "#/definitions/entity.DeckWithReviewCards"
                }
            }
        },
        "dto.CreateCardRequest": {
            "type": "object",
            "required": [
                "answer",
                "deck_id",
                "question",
                "wrong_answers"
            ],
            "properties": {
                "answer": {
                    "type": "string"
                },
                "deck_id": {
                    "type": "string"
                },
                "index": {
                    "type": "integer"
                },
                "question": {
                    "type": "string"
                },
                "question_img_label": {
                    "type": "string"
                },
                "question_img_url": {
                    "type": "string"
                },
                "wrong_answers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "dto.CreateCardResponse": {
            "type": "object",
            "properties": {
                "card": {
                    "$ref": "#/definitions/entity.Card"
                }
            }
        },
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "entity.Card": {
            "type": "object",
            "properties": {
                "answer": {
                    "type": "string"
                },
                "card_type": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "deck_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "index": {
                    "type": "integer"
                },
                "last_review": {
                    "type": "string"
                },
                "next_review": {
                    "type": "string"
                },
                "num_reviews": {
                    "type": "integer"
                },
                "question": {
                    "type": "string"
                },
                "question_img_label": {
                    "type": "string"
                },
                "question_img_url": {
                    "type": "string"
                },
                "sm2_ef": {
                    "type": "number"
                },
                "sm2_i": {
                    "type": "integer"
                },
                "sm2_n": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "string"
                },
                "wrong_answers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "entity.DeckWithCards": {
            "type": "object",
            "properties": {
                "cards": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Card"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "cur_new_cards": {
                    "type": "integer"
                },
                "cur_review_cards": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "description_img_url": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_favorite": {
                    "type": "boolean"
                },
                "is_public": {
                    "type": "boolean"
                },
                "last_review": {
                    "type": "string"
                },
                "max_new_cards": {
                    "type": "integer"
                },
                "max_review_cards": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "num_blue_cards": {
                    "type": "integer"
                },
                "num_green_cards": {
                    "type": "integer"
                },
                "num_red_cards": {
                    "type": "integer"
                },
                "position": {
                    "type": "string"
                },
                "rating": {
                    "type": "number"
                },
                "total_cards": {
                    "type": "integer"
                },
                "total_learned_cards": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "string"
                },
                "views": {
                    "type": "integer"
                }
            }
        },
        "entity.DeckWithReviewCards": {
            "type": "object",
            "properties": {
                "cards": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Card"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "cur_new_cards": {
                    "type": "integer"
                },
                "cur_review_cards": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "description_img_url": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_favorite": {
                    "type": "boolean"
                },
                "is_public": {
                    "type": "boolean"
                },
                "last_review": {
                    "type": "string"
                },
                "max_new_cards": {
                    "type": "integer"
                },
                "max_review_cards": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "num_blue_cards": {
                    "type": "integer"
                },
                "num_green_cards": {
                    "type": "integer"
                },
                "num_red_cards": {
                    "type": "integer"
                },
                "position": {
                    "type": "string"
                },
                "rating": {
                    "type": "number"
                },
                "total_cards": {
                    "type": "integer"
                },
                "total_learned_cards": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "string"
                },
                "views": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Description for what is this security definition being used",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}